requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==5.2.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
baidu-aip==4.16.8
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.7.0
brotlipy==0.7.0
cachetools==5.3.2
certifi==2021.10.8
chardet==5.0.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
easyocr==1.6.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
fire==0.4.0
flatbuffers==22.11.23
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.31
humanfriendly==10.0
imageio==2.22.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lightning-utilities==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.2.1
multidict==6.0.2
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8.8
ninja==1.11.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.3
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
openpyxl==3.0.10
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.4
pure-eval==0.2.2
pyarrow==14.0.2
pyclipper==1.3.0.post4
pycosat==0.6.3
pydeck==0.8.1b0
Pygments==2.14.0
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.8.3.post0
pytz==2022.1
PyWavelets==1.4.1
pywin32==302
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
scikit-image==0.19.3
scipy==1.9.3
Send2Trash==1.8.0
sentry-sdk==1.25.1
setproctitle==1.3.2
Shapely==1.8.5.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
streamlit==1.30.0
sympy==1.11.1
tenacity==8.2.3
tensorboardX==2.5.1
termcolor==2.1.1
terminado==0.17.1
tifffile==2022.10.10
tiffile==2018.10.18
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchaudio==0.12.1
torchmetrics==0.10.3
torchvision==0.14.1+cu117
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.4
tzlocal==5.2
Unidecode==1.3.6
uri-template==1.2.0
urllib3==1.26.16
validators==0.22.0
wandb==0.15.4
wincertstore==0.2
xlrd==1.2.0
yarl==1.8.1
zipp==3.15.0
